Move realization_state_enum
RealizationStateEnum is used several places and to
avoid circular imports when moving the config module
it needs to be at the top level.
