+++ src/ert/_c_wrappers/enkf/enums/__init__.py
@@ -1,9 +1,7 @@
 from .enkf_obs_impl_type_enum import EnkfObservationImplementationType
 from .hook_runtime_enum import HookRuntime
-from .realization_state_enum import RealizationStateEnum
 
 __all__ = [
     "EnkfObservationImplementationType",
-    "RealizationStateEnum",
     "HookRuntime",
 ]
