+++ src/ert/libres_facade.py
@@ -11,15 +11,13 @@ from ecl.grid import EclGrid
 from pandas import DataFrame, Series
 
 from ert._c_wrappers.enkf import EnKFMain, EnsembleConfig, ErtConfig
-from ert._c_wrappers.enkf.enums import (
-    EnkfObservationImplementationType,
-    RealizationStateEnum,
-)
+from ert._c_wrappers.enkf.enums import EnkfObservationImplementationType
 from ert.analysis import ESUpdate, SmootherSnapshot
 from ert.analysis._es_update import ProgressCallback
 from ert.config import Field, GenKwConfig, SurfaceConfig
 from ert.data import MeasuredData
 from ert.data._measured_data import ResponseError
+from ert.realization_state import RealizationStateEnum
 from ert.shared.version import __version__
 from ert.storage import EnsembleReader
