+++ src/ert/run_models/ensemble_experiment.py
@@ -8,8 +8,9 @@ from uuid import UUID
 
 import _ert_com_protocol
 from ert._c_wrappers.enkf import RunContext
-from ert._c_wrappers.enkf.enums import HookRuntime, RealizationStateEnum
+from ert._c_wrappers.enkf.enums import HookRuntime
 from ert.ensemble_evaluator import EvaluatorServerConfig
+from ert.realization_state import RealizationStateEnum
 from ert.storage import EnsembleAccessor, StorageAccessor
 
 from .base_run_model import BaseRunModel, ErtRunError
