+++ src/ert/run_models/iterated_ensemble_smoother.py
@@ -8,9 +8,10 @@ from iterative_ensemble_smoother import SIES
 
 from ert._c_wrappers.analysis.analysis_module import AnalysisModule
 from ert._c_wrappers.enkf import RunContext
-from ert._c_wrappers.enkf.enums import HookRuntime, RealizationStateEnum
+from ert._c_wrappers.enkf.enums import HookRuntime
 from ert.analysis import ErtAnalysisError
 from ert.ensemble_evaluator import EvaluatorServerConfig
+from ert.realization_state import RealizationStateEnum
 from ert.storage import EnsembleAccessor, StorageAccessor
 
 from .base_run_model import BaseRunModel, ErtRunError
