+++ src/ert/run_models/multiple_data_assimilation.py
@@ -4,9 +4,10 @@ import logging
 from typing import TYPE_CHECKING, Any, Dict, List, Optional
 from uuid import UUID
 
-from ert._c_wrappers.enkf.enums import HookRuntime, RealizationStateEnum
+from ert._c_wrappers.enkf.enums import HookRuntime
 from ert.analysis import ErtAnalysisError
 from ert.ensemble_evaluator import EvaluatorServerConfig
+from ert.realization_state import RealizationStateEnum
 from ert.storage import EnsembleAccessor, StorageAccessor
 
 from .base_run_model import BaseRunModel, ErtRunError
