+++ src/ert/shared/share/ert/workflows/jobs/internal-gui/scripts/gen_data_rft_export.py
@@ -6,15 +6,13 @@ import numpy
 import pandas
 
 from ert import LibresFacade
-from ert._c_wrappers.enkf.enums import (
-    EnkfObservationImplementationType,
-    RealizationStateEnum,
-)
+from ert._c_wrappers.enkf.enums import EnkfObservationImplementationType
 from ert.gui.ertwidgets.customdialog import CustomDialog
 from ert.gui.ertwidgets.listeditbox import ListEditBox
 from ert.gui.ertwidgets.models.path_model import PathModel
 from ert.gui.ertwidgets.pathchooser import PathChooser
 from ert.job_queue import CancelPluginException, ErtPlugin
+from ert.realization_state import RealizationStateEnum
 
 try:
     from PyQt4.QtGui import QCheckBox
