+++ tests/unit_tests/gui/tools/test_case_tool.py
@@ -3,11 +3,11 @@ from qtpy.QtCore import Qt
 from qtpy.QtWidgets import QPushButton, QTextEdit
 
 from ert._c_wrappers.enkf import EnKFMain, ErtConfig
-from ert._c_wrappers.enkf.enums import RealizationStateEnum
 from ert.gui.ertnotifier import ErtNotifier
 from ert.gui.tools.manage_cases.case_init_configuration import (
     CaseInitializationConfigurationPanel,
 )
+from ert.realization_state import RealizationStateEnum
 
 
 @pytest.mark.usefixtures("copy_poly_case")
